pyproject.toml
[project]
name = "deep-researcher"
version = "0.1.0"
description = "Research-to-knowledge pipeline: iterative web search, LLM summarization, and optional OpenViking ingestion"
requires-python = ">=3.10"
dependencies = [
    "litellm",
    "tavily-python",
    "httpx>=0.28.1",
    "fastmcp[tasks]>=3.2.0",
    "python-dotenv",
]
